+++ src/main/java/dave/command/AddCommand.java
@@ -1,3 +1,9 @@
+package dave.command;
+
+import dave.task.Task;
+import dave.task.TaskList;
+import dave.storage.Storage;
+import dave.ui.Ui;
 import java.io.IOException;
 
 public class AddCommand extends Command {
@@ -8,13 +14,13 @@ public class AddCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Storage storage, Ui ui) {
+    public void execute(TaskList tasks, Storage storage, Ui ui) throws IOException {
         try {
             tasks.addTask(task);
-            System.out.println("Got it. I've added this task:");
+            System.out.println("Got it. I've added this task: ");
             System.out.println(task);
             System.out.println("Now you have " + tasks.getSize() + " tasks in the list.");
-            storage.amendFile(task);
+            storage.amendFile(task);  // Ensure amendFile is accessed via the correct instance
         } catch (IOException e) {
             ui.showLine();
             System.out.println("An error occurred while saving the task to the file.");
