+++ src/main/java/dave/exceptions/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package dave.exceptions;
+
 public class InvalidCommandException extends Exception
 {
     public InvalidCommandException(String message) {
