+++ src/main/java/dave/exceptions/InvalidDescriptionException.java
@@ -1,3 +1,5 @@
+package dave.exceptions;
+
 public class InvalidDescriptionException extends Exception
 {
     public InvalidDescriptionException(String message) {
