+++ src/main/java/dave/storage/Storage.java
@@ -1,3 +1,8 @@
+package dave.storage;
+
+import dave.task.Task;
+import dave.task.TaskList;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
