+++ src/main/java/dave/task/Event.java
@@ -1,7 +1,12 @@
+package dave.task;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+import dave.exceptions.InvalidDateTimeFormatException;
+import dave.exceptions.InvalidDescriptionException;
+
 
 public class Event extends Task {
     public LocalDate fromDate;
