+++ src/main/java/dave/task/Task.java
@@ -1,3 +1,5 @@
+package dave.task;
+
 import java.io.IOException;
 
 public abstract class Task
