+++ src/main/java/dave/ui/Ui.java
@@ -1,3 +1,5 @@
+package dave.ui;
+
 import java.util.Scanner;
 
 public class Ui {
